feat: rename Checkox Component
