+++ packages/components/src/components/index.ts
@@ -4,7 +4,7 @@ export { default as Breadcrumb } from './atoms/breadcrumb/Breadcrumb'
 export { default as Breadcrumbs } from './atoms/breadcrumb/Breadcrumbs'
 export { default as Chip } from './atoms/Chip/Chip'
 export { default as CloseComponent } from './atoms/closeComponent/CloseComponent'
-export { default as Checkbox } from './atoms/form/Checkbox'
+export { default as Checkbox } from './atoms/form/CheckboxWithoutError'
 export { default as DatePicker } from './atoms/form/DatePicker'
 export { default as DatePickerError } from './atoms/form/DatePickerError'
 export { default as DoubleDatePicker } from './atoms/form/DoubleDatePicker'
